+++ game/events.py
@@ -19,6 +19,8 @@ class GameEventType(str, Enum):
     BUILDING_DESTROYED = "building_destroyed"
     BOUNTY_PLACED = "bounty_placed"
     BOUNTY_CLAIMED = "bounty_claimed"
+    HERO_ENTERED_BUILDING = "hero_entered_building"
+    HERO_EXITED_BUILDING = "hero_exited_building"
 
 
 class EventBus: